requirements.txt
ultralytics
opencv-python
pillow
requests
